[#24] Add hooks to the summoning process
